Replace individual test macros with a single one
All the (except for the with_mpxpypz one) were exactly the same thing
except for the method that was tested (and a square in case of
pt_square). By pulling the square out and passing the method to test in
as a paramter, all three macros have been replaced by a single one that
can also be used for future tests.